requirements.txt
streamlit
time
google.generativeai